requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
jinja2==3.1.4
python-multipart==0.0.9
sqlalchemy==2.0.35
python-jose[cryptography]==3.3.0
bcrypt==4.2.0
httpx==0.27.2
python-dotenv==1.0.1
psycopg2-binary==2.9.9
